python-app/requirements.txt
opencv-python==4.11.0.86
numpy==2.2.4
aiohttp==3.11.16
asyncio==3.4.3
python-dotenv==1.1.0
screeninfo==0.8.1